% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AovBay-package.R
\docType{package}
\name{AovBay-package}
\alias{AovBay-package}
\alias{AovBay}
\title{The 'AovBay' package.}
\description{
Package developed for the visualization and presentation of one-way analysis of variance models, with a classical, non-parametric and Bayesian approach.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
