% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PollutionData}
\alias{PollutionData}
\title{Pollutions Data Set}
\format{
A data frame:
\describe{
  \item{CONC.ppm}{Concentration of the pollution in parts per million.}
  \item{RemocionPorc}{Remotion Percent.}
}
}
\source{
<http://revistabionatura.com/2021.06.01.7.html>
}
\usage{
PollutionData
}
\description{
A data set of removal of a pharmaceutical product classified as emerging pollutants in aqueous medium using the vetiver species (Chrysopogon zizanioides).
}
\keyword{datasets}
